Only do work if the player is a priest; otherwise, do nothing.
Optimize by only registering for events and doing event processing
if the player is a priest, otherwise do nothing.  Leave all the
options available for setting regardless of class so that if the
player is using a shared default profile, the options for Weakened
Soul can still be set if not a priest.